+++ CHANGELOG.md
@@ -5,9 +5,13 @@
 ### Added
  - PartitionElements utility
  - Version printing callbacks for CLI
+ - Ssize utility
 
 ### Fixed
  - ReadId from CCS read names
+
+### Removed
+ - Unused SystemInfo utility
 
 ## [1.6.0] - 2020-04-10